move EndOfBuildAction
